Cleaned extraneous Auto align commands
